Clean up files and render compare layers
